requirements.txt
scrapy>=2.11
psycopg2-binary>=2.9
python-dotenv>=1.0
pandas>=2.0
